Rework newSpawnTable command
If a table already exists, it doesn't need to be re-created.  This will allow addons to define their own tables in case they're the only addons that need those tables to be loaded, but if another addon already creates that table, don't create a duplicate.
Because adding a new spawn replacement table implies a "% chance to replace this thing", being able to insert an initial row of the original thing being replaced with a default weight of 1.0 will help reduce excessive command verbosity.