Remove NullSpawner check for Replacements
It actually seems to cause a bit of an issue since the only way these are returned is if the X/256 chance fails, which is an intended behavior.  We'll see if we want to add a future "failure to replace" state that we can use to fallback to vanilla behavior or not.