Refine Spawn Table Equality Check Logic
The way it worked up until now, any two tables that shared spawnName & replacement flag would be equal, even if their unique table names were not.  This may result in a table unintentionally being selected if given the spawnName for a table with a unique name for that spawnName.

Now it picks the same prioritized property from both and compares those, along with the replacement flag.