Initial Commit to attempt seeded random
Currently, there are a few supported seeded modes: user-entered string, map checksum, and daily.  These seeds currently re-sync every map, so leaving and returning to a map via changemap or some other command won't cause the same map with the same seed to spawn things differently.  Still WIP, but the two CVARs can allow users to test and mess with the system as it gets more fleshed out.

This does introduce a new helper method, which will abstract away the currently set seed mode from the invoking code/addon.  The hope is that more things will leverage this to give the ability for players to have more consistent runs/games