Add safety checks for loading savegames and tracking stats.
Seems reopening a map pre-spawns everything that was already there when the game was saved, before the stat handlers are fully loaded/initialized.  It's fine if we ignore cases where the stats map doesn't exist though, as the handlers should run first normally and in this case we don't want to count the re-spawning of an actor a second time.  Initial tests of saving, changing maps, and reloading the save seem to reset the counters to what they were, although I'm sure more extensive testing is warranted.