Refactor out Generic Weighted Random Tables from Spawn Tables
leveraging the power of Function Pointers, the logic at which an entry is valid to be considered will now need to be provided, along with a simple dictionary of parameters that can be checked within said function.  In the Spawn Table's case, the isPersistent check, along with some other Spawn Table specific checks, will now be handled in the static HDCoreSpawnTable.shouldGetEntry(entry, params) method.  If no predicate is passed, it is assumed all entries are valid.