Refactor sector-spawner checks
It's not perfect, but the check shouldn't have gone there to begin with, it needs to check the replacement each time it attempts to spawn, not at the start, otherwise they either all spawn or none of them do.  If the option is set to disabled, it'll skip the entire method.